# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/donations": {
      "get": {
        "operationId": "getDonations",
        "summary": "List donation records with pagination.",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 25,
              "example": 25,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DonationsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Donations"
        ]
      },
      "post": {
        "operationId": "DonationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDonationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DonationDto"
                }
              }
            }
          }
        },
        "tags": [
          "Donations"
        ]
      }
    },
    "/donations/{id}/status": {
      "patch": {
        "operationId": "DonationsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDonationStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DonationDto"
                }
              }
            }
          }
        },
        "tags": [
          "Donations"
        ]
      }
    },
    "/donations/webhooks/paystack": {
      "post": {
        "operationId": "DonationsController_handlePaystackWebhook",
        "parameters": [
          {
            "name": "x-paystack-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Donations"
        ]
      }
    },
    "/donations/webhooks/fincra": {
      "post": {
        "operationId": "DonationsController_handleFincraWebhook",
        "parameters": [
          {
            "name": "x-fincra-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Donations"
        ]
      }
    },
    "/donations/webhooks/stripe": {
      "post": {
        "operationId": "DonationsController_handleStripeWebhook",
        "parameters": [
          {
            "name": "stripe-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Donations"
        ]
      }
    },
    "/donations/webhooks/flutterwave": {
      "post": {
        "operationId": "DonationsController_handleFlutterwaveWebhook",
        "parameters": [
          {
            "name": "verif-hash",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Donations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Covenant Connect API",
    "description": "OpenAPI documentation for the Covenant Connect backend services.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "DonationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "don-1"
          },
          "memberId": {
            "type": "string",
            "nullable": true,
            "example": "member-1"
          },
          "amount": {
            "type": "number",
            "example": 125
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "provider": {
            "type": "string",
            "enum": [
              "paystack",
              "fincra",
              "stripe",
              "flutterwave"
            ],
            "example": "stripe"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "completed",
              "failed",
              "refunded"
            ],
            "example": "completed"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "memberId",
          "amount",
          "currency",
          "provider",
          "status",
          "metadata",
          "createdAt",
          "updatedAt"
        ]
      },
      "DonationsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DonationDto"
            }
          },
          "total": {
            "type": "number",
            "example": 120
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 25
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "pageSize"
        ]
      },
      "CreateDonationDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateDonationStatusDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}